+++ tests/end_to_end/target_snowflake/tap_mysql/__init__.py
@@ -5,6 +5,7 @@ MYSQL_FASTSYNC_TABLES = {
     'iceberg_events': True,
     'iceberg_incremental': True,
     'iceberg_full_reload': False,
+    'multiline_values': True,
 }